# openapi/auth-user-service.openapi.yaml
openapi: 3.1.0
info:
  title: SmartCloud-X Auth User Service API
  version: 0.1.0
  description: Foundation-owned placeholder baseline for user auth, admin auth bootstrap, and internal permission-check routes aligned to spec sections 20.13, 20.14, and 20.17.
...
servers:
  - url: /
paths:
...
  /api/v1/auth/login:
    post:
      tags: [user-auth]
      operationId: auth_login
      summary: Login for the current end-user surface.
      x-owner-service: auth-user-service
      x-permission-code: anonymous
      x-error-codes: [4001001, 4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: public-auth
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./components.openapi.yaml#/components/schemas/LoginRequest
      responses:
        '200':
          description: User login success response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessEnvelope'
              example:
                code: 0
                message: ok
                request_id: req-login-1
                timestamp: 1776300000000
                data:
                  access_token: access-token
                  refresh_token: refresh-token
                  expires_in: 7200
                  user:
                    user_id: user-demo
                    tenant_id: tenant-demo
                    name: Demo User
                    email: demo@smartcloud.local
                    mobile: '13800138000'
                    avatar_url: null
                    locale: zh-CN
                    time_zone: Asia/Shanghai
                    permissions: [user:chat.use, user:research.read]
        '400':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalValidationError
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
  /api/v1/auth/send-code:
    post:
      tags: [user-auth]
      operationId: auth_sendCode
      summary: Send a login or password-reset verification code.
      x-owner-service: auth-user-service
      x-permission-code: anonymous
      x-error-codes: [4001001, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: public-auth
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./components.openapi.yaml#/components/schemas/SendCodeRequest
      responses:
        '200':
          description: Verification code accepted for delivery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendCodeSuccessEnvelope'
        '400':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalValidationError
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
  /api/v1/auth/refresh:
    post:
      tags: [user-auth]
      operationId: auth_refresh
      summary: Refresh the user access token pair.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:user
      x-error-codes: [4001001, 4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: public-auth
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./components.openapi.yaml#/components/schemas/RefreshTokenRequest
      responses:
        '200':
          description: Token refresh response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenSuccessEnvelope'
        '400':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalValidationError
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
  /api/v1/auth/me:
    get:
      tags: [user-auth]
      operationId: auth_getCurrentUser
      summary: Return the current authenticated user profile and permission set.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:user
      x-error-codes: [4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: user
      security:
        - BearerAuth: []
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      responses:
        '200':
          description: Current user profile response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserSuccessEnvelope'
              example:
                code: 0
                message: ok
                request_id: req-profile-1
                timestamp: 1776300000000
                data:
                  user_id: user-demo
                  tenant_id: tenant-demo
                  name: Demo User
                  email: demo@smartcloud.local
                  mobile: '13800138000'
                  avatar_url: null
                  locale: zh-CN
                  time_zone: Asia/Shanghai
                  permissions: [user:chat.use, user:research.read]
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
  /api/v1/auth/profile:
    get:
      tags: [user-profile]
      operationId: auth_getCurrentUserProfileAlias
      summary: Return the current authenticated user profile through the mission-wording compatibility alias.
      description: Additive compatibility alias for the same profile payload served by `/api/v1/auth/me`.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:user
      x-error-codes: [4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: user
      security:
        - BearerAuth: []
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      responses:
        '200':
          description: Current user profile response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserSuccessEnvelope'
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
    patch:
      tags: [user-profile]
      operationId: auth_updateCurrentUserProfileAlias
      summary: Update the mutable fields on the current user profile through the mission-wording compatibility alias.
      description: Additive compatibility alias for the same profile update contract served by `/api/v1/users/me`.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:user
      x-error-codes: [4001001, 4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: user
      security:
        - BearerAuth: []
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./components.openapi.yaml#/components/schemas/UserProfileUpdateRequest
      responses:
        '200':
          description: Updated current user profile response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserSuccessEnvelope'
        '400':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalValidationError
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
  /api/v1/auth/logout:
    post:
      tags: [user-auth]
      operationId: auth_logout
      summary: Logout the current user and revoke the refresh token when provided.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:user
      x-error-codes: [4001001, 4010002, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: user
      security:
        - BearerAuth: []
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: ./components.openapi.yaml#/components/schemas/LogoutRequest
      responses:
        '200':
          description: Logout succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatusSuccessEnvelope'
        '400':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalValidationError
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
...
  /api/v1/admin/auth/me:
    get:
      tags: [admin-auth]
      operationId: adminAuth_getCurrentAdmin
      summary: Return the current authenticated admin profile, roles, permissions, and menus.
      x-owner-service: auth-user-service
      x-permission-code: authenticated:admin
      x-error-codes: [4010001, 4290001]
      x-idempotency-required: false
      x-rate-limit-scope: admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: ./components.openapi.yaml#/components/parameters/XRequestIdHeader
      responses:
        '200':
          description: Current admin profile response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionSuccessEnvelope'
        '401':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalUnauthorized
        '429':
          $ref: ./components.openapi.yaml#/components/responses/CanonicalRateLimited
...
components:
  schemas:
    LoginSuccessEnvelope:
      allOf:
        - $ref: ./components.openapi.yaml#/components/schemas/CanonicalSuccessEnvelope
        - type: object
          properties:
            data:
              $ref: ./components.openapi.yaml#/components/schemas/LoginResponseData
    SendCodeSuccessEnvelope:
      allOf:
        - $ref: ./components.openapi.yaml#/components/schemas/CanonicalSuccessEnvelope
        - type: object
          properties:
            data:
              $ref: ./components.openapi.yaml#/components/schemas/SendCodeResponseData
    RefreshTokenSuccessEnvelope:
      allOf:
        - $ref: ./components.openapi.yaml#/components/schemas/CanonicalSuccessEnvelope
        - type: object
          properties:
            data:
              $ref: ./components.openapi.yaml#/components/schemas/RefreshTokenResponseData
    CurrentUserSuccessEnvelope:
      allOf:
        - $ref: ./components.openapi.yaml#/components/schemas/CanonicalSuccessEnvelope
        - type: object
          properties:
            data:
              $ref: ./components.openapi.yaml#/components/schemas/AuthUserProfile
...
    OperationStatusSuccessEnvelope:
      allOf:
        - $ref: ./components.openapi.yaml#/components/schemas/CanonicalSuccessEnvelope
        - type: object
          properties:
            data:
              $ref: ./components.openapi.yaml#/components/schemas/OperationStatusData
...